Car Warranty
check car warranty status, test by subtracting car year - current year. Test if the amount of years is greater than 5 OR if the milage is greater than 60,000. Return true if it is valid and false if it is not. Then push the result to a new array. Return the results to the screen.